♻️ refactor: renderField 함수 통합 #330
