+++ src/components/Employer/WriteDocument/EmployerLaborContractForm.tsx
@@ -10,21 +10,14 @@ import {
   LaborContractEmployerInfo,
   Phone,
 } from '@/types/api/document';
-import { InputType } from '@/types/common/input';
-import SignaturePad from '@/components/Document/write/SignaturePad';
 import BottomButtonPanel from '@/components/Common/BottomButtonPanel';
 import { usePutLaborContractEmployer } from '@/hooks/api/useDocument';
 import { validateLaborContractEmployerInformation } from '@/utils/document';
 import { formatPhoneNumber, parsePhoneNumber } from '@/utils/information';
 import { useParams } from 'react-router-dom';
-import { Controller, useForm } from 'react-hook-form';
+import { useForm } from 'react-hook-form';
 import ValidatedSubmitButton from '@/components/Document/write/ValidatedSubmitButton';
-import DocumentFormInput from '@/components/Document/write/input/DocumentFormInput';
-import PhoneNumberInput from '@/components/Document/write/input/PhoneNumberInput';
-import AddressInput from '@/components/Document/write/input/AddressInput';
-import RadioGroup from '@/components/Document/write/input/RadioGroup';
-import CheckboxGroup from '@/components/Document/write/input/CheckboxGroup';
-import WorkDayTimeWithRestInput from '@/components/Document/write/input/WorkDayTimeWithRestInput';
+import { renderField } from '@/components/Document/write/renderField';
 
 type LaborContractFormProps = {
   document?: LaborContractDataResponse;
@@ -84,139 +77,13 @@ const EmployerLaborContractForm = ({
   };
 
   // 폼 필드 렌더링 함수
-  const renderField = (field: LaborContractEmployerFormField) => {
-    switch (field.type) {
-      case 'text':
-        return (
-          <DocumentFormInput
-            inputType={InputType.TEXT}
-            placeholder={field.placeholder}
-            canDelete={false}
-            name={field.name as keyof LaborContractEmployerInfo}
-            control={control}
-            format={field.format}
-            description={field.description}
-            isUnit={field.isUnit}
-            unit={field.unit}
-            isPrefix={field.isPrefix}
-            prefix={field.prefix}
-          />
-        );
-      case 'textarea':
-        return (
-          <Controller
-            control={control}
-            name={field.name as keyof LaborContractEmployerInfo}
-            render={({ field: { value, onChange } }) => (
-              <div className="w-full self-stretch flex flex-col items-center justify-start body-3">
-                <div className="w-full flex flex-col items-start justify-start">
-                  <div className="w-full flex flex-col items-center justify-start">
-                    <textarea
-                      className="w-full h-[10vh] px-[1rem] py-[0.75rem] border border-border-alternative rounded-[0.75rem] body-2 outline-none resize-none"
-                      placeholder={field.placeholder}
-                      value={value as string}
-                      onChange={onChange}
-                    />
-                  </div>
-                </div>
-              </div>
-            )}
-          />
-        );
-      case 'phone':
-        return <PhoneNumberInput control={control} name={field.name} />;
-      case 'address':
-        return (
-          <AddressInput
-            control={control}
-            name={field.name as string}
-            placeholder={field.placeholder}
-            label={field.label}
-          />
-        );
-      case 'work_schedule':
-        return (
-          <WorkDayTimeWithRestInput
-            control={control}
-            name={field.name as keyof LaborContractEmployerInfo}
-          />
-        );
-      case 'signature':
-        return (
-          <Controller
-            control={control}
-            name={field.name as keyof LaborContractEmployerInfo}
-            render={({ field: { value, onChange } }) => (
-              <div
-                className={`w-full relative shadow rounded-xl box-border h-[120px] mb-40`}
-              >
-                <SignaturePad
-                  onSave={onChange}
-                  onReset={() => onChange('')}
-                  previewImg={value as string}
-                />
-              </div>
-            )}
-          />
-        );
-
-      case 'radio': {
-        return (
-          <RadioGroup
-            control={control}
-            name={field.name}
-            options={field.options || []}
-            description={field.description}
-            transformer={field.transformer}
-          />
-        );
-      }
-      case 'checkbox': {
-        return (
-          <CheckboxGroup
-            control={control}
-            name={field.name}
-            options={field.checkboxOptions || []}
-            description={field.description}
-            variant={field.variant}
-          />
-        );
-      }
-      case 'input_with_radio': {
-        return (
-          <div>
-            <Controller
-              control={control}
-              name={field.name as keyof LaborContractEmployerInfo}
-              render={({ field: { value } }) => (
-                <div className="flex flex-col gap-2">
-                  <RadioGroup
-                    control={control}
-                    name={field.name}
-                    options={field.options || []}
-                    description={field.description}
-                    transformer={field.transformer}
-                  />
-                  {value === 0 && (
-                    <DocumentFormInput
-                      inputType={InputType.TEXT}
-                      placeholder={field.placeholder}
-                      canDelete={false}
-                      name={field.name as keyof LaborContractEmployerInfo}
-                      control={control}
-                      isUnit={field.isUnit}
-                      unit={field.unit}
-                    />
-                  )}
-                </div>
-              )}
-            />
-          </div>
-        );
-      }
-      default:
-        return null;
-    }
+  // 폼 필드 렌더링 함수
+  const renderFormField = (field: LaborContractEmployerFormField) => {
+    return renderField<LaborContractEmployerInfo>({
+      field,
+      control,
+      name: field.name as keyof LaborContractEmployerInfo,
+    });
   };
 
   return (
@@ -228,7 +95,7 @@ const EmployerLaborContractForm = ({
           {/** 고용주 작성 정보 */}
           {LaborContractEmployerFormFields.map((field) => (
             <InputLayout key={field.name} title={field.title} isEssential>
-              {renderField(field)}
+              {renderFormField(field)}
             </InputLayout>
           ))}
         </div>
