+++ src/components/WriteDocuments/LaborContractWriteForm.tsx
@@ -10,10 +10,8 @@ import {
   Phone,
 } from '@/types/api/document';
 import { formatPhoneNumber, parsePhoneNumber } from '@/utils/information';
-import { InputType } from '@/types/common/input';
 import { validateLaborContract } from '@/utils/document';
 import BottomButtonPanel from '@/components/Common/BottomButtonPanel';
-import SignaturePad from '@/components/Document/write/SignaturePad';
 import EmployerInfoSection from '@/components/Document/write/EmployerInfoSection';
 import {
   usePostStandardLaborContracts,
@@ -21,11 +19,9 @@ import {
 } from '@/hooks/api/useDocument';
 import InputLayout from '@/components/WorkExperience/InputLayout';
 import { useParams } from 'react-router-dom';
-import { useForm, Controller } from 'react-hook-form';
-import DocumentFormInput from '@/components/Document/write/input/DocumentFormInput';
-import PhoneNumberInput from '@/components/Document/write/input/PhoneNumberInput';
+import { useForm } from 'react-hook-form';
 import ValidatedSubmitButton from '@/components/Document/write/ValidatedSubmitButton';
-import AddressInput from '@/components/Document/write/input/AddressInput';
+import { renderField } from '@/components/Document/write/renderField';
 
 // 필수 검증 필드 목록
 const REQUIRED_FIELDS: Array<keyof LaborContractEmployeeInfo | 'phone'> = [
@@ -97,50 +93,12 @@ const LaborContractWriteForm = ({
   };
 
   // 필드 타입에 따른 입력 컴포넌트 렌더링
-  const renderField = (field: LaborContractFormField) => {
-    switch (field.type) {
-      case 'text':
-        return (
-          <DocumentFormInput
-            inputType={InputType.TEXT}
-            placeholder={field.placeholder}
-            canDelete={false}
-            name={field.name as keyof LaborContractEmployeeInfo}
-            control={control}
-            format={field.format}
-          />
-        );
-      case 'phone':
-        return <PhoneNumberInput control={control} name={field.name} />;
-      case 'address':
-        return (
-          <AddressInput
-            control={control}
-            name={field.name as string}
-            placeholder={field.placeholder}
-          />
-        );
-      case 'signature':
-        return (
-          <Controller
-            control={control}
-            name="signature_base64"
-            render={({ field: { value, onChange } }) => (
-              <div
-                className={`w-full relative shadow rounded-xl box-border h-[120px] mb-40`}
-              >
-                <SignaturePad
-                  onSave={onChange}
-                  onReset={() => onChange('')}
-                  previewImg={value || ''}
-                />
-              </div>
-            )}
-          />
-        );
-      default:
-        return null;
-    }
+  const renderFormField = (field: LaborContractFormField) => {
+    return renderField<LaborContractEmployeeInfo>({
+      field,
+      control,
+      name: field.name as keyof LaborContractEmployeeInfo,
+    });
   };
 
   // 폼이 비활성화되어야 하는지 여부
@@ -156,7 +114,7 @@ const LaborContractWriteForm = ({
           {/* 유학생 작성 정보 */}
           {LaborContractFormFields.map((field) => (
             <InputLayout key={field.name} title={field.title} isEssential>
-              {renderField(field)}
+              {renderFormField(field)}
             </InputLayout>
           ))}
           {/* 고용주 정보가 있다면 표시 */}
